Refactor: decompose typechecker; diagnostics echo the user's spelling
Two related changes, reviewed together:

1) Decompose the ~1500-line `flowlog-typechecker/src/lib.rs` along the
   two-pass structure it already had. The passes stay separate (they reason
   in different type lattices — primitive `DataType` + literal pinning vs
   nominal `TypeId` + `as()` cast lowering — and encode real diagnostic
   ordering and mutation timing; fusing was evaluated and rejected).

   - `lib.rs`     — `check_program` orchestration only; reads as "Pass 1; Pass 2".
   - `env.rs`     — `PrimitiveEnv { decls, udfs }` built once per program.
   - `primitive/` — Pass 1: `mod` (entry points), `lattice` (`LitKind`
                    + `of`/`merge`), `expr` (inference + pinning),
                    `check` (per-rule walk).
   - `subtype.rs` — Pass 2, kept whole.
   - Inline test block split into `tests/{pinning,subtype,tuple}.rs` with a
     shared `tests/common` helper.
   - Call style: pass drivers take `&PrimitiveEnv`; leaf helpers take
     exactly the map they use. Entry points named in parallel
     (`check_and_pin_rules` / `check_and_pin_facts` / `check_and_lower`).
   - `TypeCheckError` is `#[non_exhaustive]` (published-crate insurance:
     future checks can add variants without a semver break).

2) Type errors now echo the relation name exactly as the user spelled it
   at the error site, and the canonical→raw `DisplayNames` map is gone:

   - Head errors (both passes) read `head.raw_name()` off the AST.
   - Fact errors read the new `InlineFact { span, raw_name, columns }`
     record (`flowlog-parser/src/program/fact.rs`), which replaces the
     anonymous `(Span, Vec<ConstType>)` tuple in `Program`'s facts map —
     the parser had the spelling in hand at `extract_fact` and previously
     dropped it. Consumers in build/compiler EDB codegen updated
     mechanically.

Verified: workspace tests (incl. 26 typechecker error fixtures), clippy
`--workspace --all-targets -D warnings`, `cargo fmt --check`, and the full
e2e fixture suite — 121 programs through BOTH runners (binary + lib mode),
byte-diffed against expected outputs — all green on the final tree.
